+++ explabox/explain/text/__init__.py
@@ -12,17 +12,12 @@ from ...utils import MultipleReturn
 
 
 class Explainer(Readable, IngestiblesMixin):
-    def __init__(self, data, model, labelprovider, ingestibles=None):
-        """Single object to create explanations corresponding to a model and dataset (with ground-truth labels).
-
-        Args:
-            model: ...
-            labelprovider: ...
-        """
+    def __init__(self, data=None, model=None, ingestibles=None):
+        """Single object to create explanations corresponding to a model and dataset (with ground-truth labels)."""
         if ingestibles is None:
-            ingestibles = Ingestible(data=data, model=model, labelprovider=labelprovider)
+            ingestibles = Ingestible(data=data, model=model)
         self.ingestibles = ingestibles
-        self.check_requirements(['data', 'labelprovider', 'model'])
+        self.check_requirements(['data', 'model'])
 
     @restyle
     def explain_prediction(